wpt_babel/content-security-policy/nonce-hiding/dangling-html-or-body.tentative.html

<!DOCTYPE html>
<script nonce="secret" src="/resources/testharness.js">
</script>
<script nonce="secret" src="/resources/testharnessreport.js">
</script>
<!-- `Content-Security-Policy: script-src 'nonce-secret'` delivered via headers -->
<body>
 <style>
  body[nonce*=secret]{background:url(/security/resources/abe.png);}
 </style>
 <body <script="" nonce="secret" src="https://example.com/good.js">
  <script nonce="secret">
   "use strict";

test(function (t) {
  var body = document.querySelector('body');
  var style = getComputedStyle(body);
  assert_equals(style['background-image'], 'none');
}, "Nonces don't leak via CSS side-channels when a dangling body is injected.");
  </script>
  <style>
   html[nonce*=secret]{background:url(/security/resources/abe.png);}
  </style>
  <html <script="" nonce="secret" src="https://example.com/good.js">
   <script nonce="secret">
    "use strict";

test(function (t) {
  var html = document.querySelector('html');
  var style = getComputedStyle(html);
  assert_equals(style['background-image'], 'none');
}, "Nonces don't leak via CSS side-channels when a dangling html is injected.");
   </script>
  </html>
 </body>
</body>
